port latest changes from winforms app
